autooptimize: extract shared log-trace helpers from CLI discord/telegram commands (-106 LOC)
